+++ common/logging_test.go
@@ -1,0 +1,14 @@
+package common
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestStructuredLoggingAdapterIsInitialized(t *testing.T) {
+	require.NotNil(t, Logging)
+	require.NotPanics(t, func() {
+		Logging.Infof("signal handler logging test")
+	})
+}